refactor: split src/index.ts into modules (closes #8)
Split the monolithic 1,100-line src/index.ts into logical modules:

- src/types.ts: TypeScript interfaces (ToolTextResult, CommandExecutionOptions, etc.)
- src/utils.ts: Utility functions (resolveNativeBinary, executeCommand, runNative, runSimctl, etc.) and shared constants
- src/tools/info.ts: baepsae_help, baepsae_version, list_apps, get_focused_app
- src/tools/simulator.ts: list_simulators, open_url, install_app, uninstall_app, launch_app, terminate_app
- src/tools/ui.ts: describe_ui, search_ui, tap, type_text, swipe, scroll, drag_drop
- src/tools/input.ts: key, key_sequence, key_combo, button, touch, gesture
- src/tools/media.ts: screenshot, record_video, stream_video
- src/tools/system.ts: list_windows, activate_app, screenshot_app, right_click, menu_action, clipboard

Each tool file exports a registerXxxTools(server) function. The entry point
(src/index.ts) imports and calls all registration functions. No tool behavior,
names, or schemas were changed. All 12 contract tests pass.
